changed Amount field to float and prepared for deploy

+++ mmappbackend/transactions/models.py
@@ -30,7 +30,7 @@ class TransactionType(models.Model):
 
 class Transaction(models.Model):
     id = models.UUIDField(primary_key=True,default=uuid.uuid4, editable=False)
-    amount = models.PositiveIntegerField()
+    amount = models.FloatField()
     description = models.CharField(max_length=100)
     created_date = models.DateTimeField(auto_now=True)
     modified_date = models.DateTimeField(auto_now_add=True)
